% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_party_memberships}
\alias{combine_party_memberships}
\title{Combine consecutive records in a tibble of party memberships}
\usage{
combine_party_memberships(pm)
}
\arguments{
\item{collapse}{A tibble containing party memberships as returned by one of
the fetch party membership functions.}
}
\value{
A tibble of party memberships, with one row per party membership.
The memberships are processed and combined so that there is only one party
membership for a period of continuous membership within the same party.
}
\description{
\code{combine_party_memberships} takes a tibble of party memberships and
combines historically consecutive memberships of the same party into a
single continuous memberships with the start date of thre first membership
and the end date of the last. Combining the memberships in this way means
that party membership ids from the data platform are not included in the
dataframe returned.
}
\keyword{internal}
