% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_pdp_date}
\alias{convert_pdp_date}
\title{Convert a datetype variable returned from the data platorm to an R Date}
\usage{
convert_pdp_date(datetime_str)
}
\description{
Convert a datetype variable returned from the data platorm to an R Date
}
\keyword{internal}
