% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{date_format_error}
\alias{date_format_error}
\title{Report an error parsing a date string}
\usage{
date_format_error(date_str)
}
\arguments{
\item{date_str}{The date string that could not be parsed.}
}
\description{
Report an error parsing a date string
}
\keyword{internal}
