% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{fetch_committee_memberships_raw}
\alias{fetch_committee_memberships_raw}
\title{Fetch committee memberships for Members}
\usage{
fetch_committee_memberships_raw(house = NA)
}
\description{
Fetch committee memberships for Members
}
\keyword{internal}
