% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{fetch_government_roles_raw}
\alias{fetch_government_roles_raw}
\title{Fetch government roles for Members}
\usage{
fetch_government_roles_raw(house = NA)
}
\description{
Fetch government roles for Members
}
\keyword{internal}
