% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_memberships}
\alias{fetch_lords_memberships}
\title{Fetch Lords memberships for all Lords}
\usage{
fetch_lords_memberships(from_date = NA, to_date = NA, on_date = NA)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the on_date.}
}
\value{
A tibble of Lords memberships for each Lord, with one row per
  Lords membership.
}
\description{
\code{fetch_lords_memberships} fetches data from the data platform showing
Lords memberships for each Lord.
}
\details{
The from_date and to_date arguments can be used to filter the memberships
returned. The on_date argument is a convenience that sets the from_date and
to_date to the same given date. The on_date has priority: if the on_date is
set, the from_date and to_date are ignored.

The filtering is inclusive: a membership is returned if any part of it falls
within the period specified with the from and to dates.

Note that a membership with an NA end date is still open.
}
