% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_opposition_roles}
\alias{fetch_lords_opposition_roles}
\title{Fetch opposition roles for all Lords}
\usage{
fetch_lords_opposition_roles(
  from_date = NA,
  to_date = NA,
  on_date = NA,
  while_lord = TRUE
)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the on_date.}

\item{while_lord}{A boolean indicating whether to filter the opposition
roles to include only those roles that were held while each individual was
serving as an MP. The default value is TRUE.}
}
\value{
A tibble of opposition roles for each MP, with one row per role.
}
\description{
\code{fetch_lords_oppositon_roles} fetches data from the data platform
showing opposition roles for each Lord.
}
\details{
The from_date and to_date arguments can be used to filter the roles
returned. The on_date argument is a convenience that sets the from_date and
to_date to the same given date. The on_date has priority: if the on_date is
set, the from_date and to_date are ignored.

The filtering is inclusive: a role is returned if any part of it falls
within the period specified with the from and to dates.

The while_lord argument can be used to filter the roles to include only
those that occurred during the period when each individual was a Lord.

Note that a role with an NA end date is still open.
}
