% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_opposition_roles_raw}
\alias{fetch_lords_opposition_roles_raw}
\title{Fetch opposition roles for all Lords}
\usage{
fetch_lords_opposition_roles_raw()
}
\description{
Fetch opposition roles for all Lords
}
\keyword{internal}
