% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords.R
\name{fetch_lords_party_memberships_raw}
\alias{fetch_lords_party_memberships_raw}
\title{Fetch party memberships for all Lords}
\usage{
fetch_lords_party_memberships_raw()
}
\description{
Fetch party memberships for all Lords
}
\keyword{internal}
