% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{fetch_mps_opposition_roles_raw}
\alias{fetch_mps_opposition_roles_raw}
\title{Fetch opposition roles for all MPs}
\usage{
fetch_mps_opposition_roles_raw()
}
\description{
Fetch opposition roles for all MPs
}
\keyword{internal}
