% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{fetch_mps_party_memberships_raw}
\alias{fetch_mps_party_memberships_raw}
\title{Fetch party memberships for all MPs}
\usage{
fetch_mps_party_memberships_raw()
}
\description{
Fetch party memberships for all MPs
}
\keyword{internal}
