% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{fetch_opposition_roles_raw}
\alias{fetch_opposition_roles_raw}
\title{Fetch opposition roles for Members}
\usage{
fetch_opposition_roles_raw(house = NA)
}
\description{
Fetch opposition roles for Members
}
\keyword{internal}
