% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{fetch_party_memberships_raw}
\alias{fetch_party_memberships_raw}
\title{Fetch party memberships for Members}
\usage{
fetch_party_memberships_raw(house = NA)
}
\description{
Fetch party memberships for Members
}
\keyword{internal}
