% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_dates}
\alias{filter_dates}
\title{Filter a tibble of data based on the given from and to dates.}
\usage{
filter_dates(df, start_col, end_col, from_date = NA, to_date = NA)
}
\arguments{
\item{df}{A tibble containing data on a time bound activity.}

\item{start_col}{The name of the column that contains the start date for
the activity.}

\item{end_col}{The name of the column that contains the end date for the
activity.}

\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. 2000-12-31'. The
default value is NA, which means no records are excluded on the basis
of the to_date.}
}
\value{
A tibble with the same structure as the input df containing
  the rows that meet the filtering criteria.
}
\description{
\code{filter_dates} takes a tibble which contains data on a time bound
activity and returns the subset of rows where that activity took place
within a given period. The tibble must contain two columns of Date objects,
which record the start and end dates of an activity. The from and to dates
provided are used to find all rows where some part of the period
of activity took place within the period of filtering. The filtering
process is inclusive: as long as at least one day of activity falls within
the filtering period, the row is returned.
}
\keyword{internal}
