% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get_api_url}
\alias{get_api_url}
\title{Get the api url}
\usage{
get_api_url()
}
\value{
The currently set api url as a string.
}
\description{
\code{get_api_key} gets the url that the package is currently configured to
use for the SPARQL endpoint to a data platform instance.
}
\keyword{internal}
