% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{get_general_elections}
\alias{get_general_elections}
\title{Return the dates of UK general elections since 1929 as a tibble}
\usage{
get_general_elections()
}
\value{
A tibble with data on general elections.
}
\description{
\code{get_general_elections} returns the dates of UK general elections since
1929 as a tibble with three columns:
}
\details{
name        -- The name of each general election as a string
 dissolution -- The date of dissolution as a date
 election    -- The date of the election as a date
}
