% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{get_general_elections_list}
\alias{get_general_elections_list}
\title{Return the dates of UK general elections since 1929 as a list}
\usage{
get_general_elections_list()
}
\value{
A list containing the dissolution and election dates for each
  general election.
}
\description{
\code{get_general_elections_list} returns a list containing the dissolution
and election dates for each general election since 1929 as date objects.
Each item in the list is named with the election name and contains a list of
two values: one named "dissolution" containing the dissolution date and the
other named "election" containing the election date.
}
