% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{handle_date}
\alias{handle_date}
\title{Take a date which may be a string or a date and returns a date.}
\usage{
handle_date(d)
}
\description{
\code{handle_date} takes a date which may be a Date or an ISO 8601 date
string, checks it is valid, and returns the date as a Date. NA values are
returned unmodified. This function raises an error if it is unable to
handle the date.
}
\keyword{internal}
