% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{missing_column_error}
\alias{missing_column_error}
\title{Report an error handling dataframes with missing columms}
\usage{
missing_column_error(colname)
}
\arguments{
\item{colname}{The name of the column that could not be found.}
}
\description{
Report an error handling dataframes with missing columms
}
\keyword{internal}
