% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdpr.R
\docType{package}
\name{pdpr}
\alias{pdpr}
\alias{pdpr-package}
\title{pdpr: A package for downloading data from the Parliamentary Data Platform}
\description{
The pdpr package provides a suite of functions for downloading data from
the data platform for the UK Parliament.
}
