% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readable}
\alias{readable}
\title{Take a tibble and remove all columns that end in the suffix "_id"}
\usage{
readable(df)
}
\arguments{
\item{df}{A tibble.}
}
\value{
A tibble with the same structure as the input df with any columns
ending in the suffix "_id" removed.
}
\description{
The intended purpose of this function is to display a tibble on the console
showing only the readable columns i.e. not the identifiers.
}
