% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{request}
\alias{request}
\title{Send an http request with a query and return the response}
\usage{
request(query)
}
\arguments{
\item{query}{A SPARQL query as a string.}
}
\value{
The http response object from httr.
}
\description{
\code{request} sends a SPARQL query to the api endpoint and returns the
response object. It is a simple wrapper around httr::POST. It sets the
appropriate headers and sends the query as the request body. It does not
validate the query or handle the response in any way. The response format
is JSON.
}
\keyword{internal}
