% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{reset_api_url}
\alias{reset_api_url}
\title{Reset the api url to the default}
\usage{
reset_api_url()
}
\description{
\code{reset_api_url} resets the url that the package uses for the api
endpoint to the live api url.
}
