% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_api_url}
\alias{set_api_url}
\title{Set the api url}
\usage{
set_api_url(api_url)
}
\arguments{
\item{api_url}{The url of an available data platform SPARQL endpoint.}
}
\description{
\code{set_api_url} sets the url that the package uses for the api endpoint.
By default the package uses the main live endpoint for the data platform's
SPARQL api. If you wish to run a local version of the api you can use this
function to tell the package to use that endpoint instead.
}
