% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{settings}
\alias{settings}
\title{Package settings environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
settings
}
\description{
Package settings environment
}
\keyword{internal}
