% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sparql_select}
\alias{sparql_select}
\title{Send a select query and return the response as a tibble}
\usage{
sparql_select(query)
}
\arguments{
\item{query}{A SPARQL SELECT query as a string.}
}
\value{
A tibble containing the results of the query.
}
\description{
\code{sparql_select} sends a SPARQL query to the api endpoint and returns
the response as a tibble. The SPARQL should be a SELECT query as the
response is processed as tabular data. The function will convert datatypes
that it recognises. It currently recognises date types. All other data
returned in the tibble will be strings. If the query syntax is not valid or
the request fails for any other reason an error will be raised with the
response text.
}
